Added DateTimeExtension class
